pyproject.toml
[project]
name = "shuxiang"
version = "0.0.1"
description = "Bilingual operator agent for US bureaucratic systems — hackathon demo (Illinois LLC filing, Mandarin)"
requires-python = ">=3.11,<3.14"
dependencies = [
    "browser-use>=0.1.40",
    "playwright>=1.47.0",
    "anthropic>=0.39.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "httpx>=0.27.0",
    "sounddevice>=0.5.0",
    "webrtcvad>=2.0.10",
    "numpy>=2.0.0",
    "setuptools<81",  # webrtcvad still imports pkg_resources
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "voice", "overlay", "pdf"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
